+++ CHANGELOG.md
@@ -5,6 +5,12 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix edge case, where a manual timebased toggle would not turn on the device at it would be off by it's
+  time settings
+
 ## 3.0.0-alpha.13 (2024-03-17)
 
 * (theimo1221) TargetAutomaticState might be wrong, if the automatic action fired before the manual action overwrote it.